+++ Vue_Full_Project/src/views/base/Carousels.vue
@@ -7,9 +7,9 @@
             header-tag="header"
             footer-tag="footer">
             <div slot="header">
-              <i className="fa fa-align-justify"></i><strong>Bootstrap Carousel</strong>
+              <i class="fa fa-align-justify"></i> <strong>Bootstrap Carousel</strong>
               <div class="card-actions">
-                <a href="https://bootstrap-vue.js.org/docs/components/carousel" target="_blank"><small className="text-muted">docs</small></a>
+                <a href="https://bootstrap-vue.js.org/docs/components/carousel" target="_blank"><small class="text-muted">docs</small></a>
               </div>
             </div>
             <div>