+++ Vue_Full_Project/src/views/base/ProgressBars.vue
@@ -5,10 +5,10 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Bootstrap Progress</strong>
+          <i class="fa fa-align-justify"></i> <strong>Bootstrap Progress</strong>
           <div class="card-actions">
             <a href="https://bootstrap-vue.js.org/docs/components/progress" target="_blank">
-              <small className="text-muted">docs</small>
+              <small class="text-muted">docs</small>
             </a>
           </div>
         </div>
@@ -26,7 +26,7 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Progress</strong> <small>labels</small>
+          <i class="fa fa-align-justify"></i> <strong>Progress</strong> <small>labels</small>
         </div>
         <div>
           <h5>No label</h5>
@@ -45,7 +45,7 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Progress</strong> <small>width</small>
+          <i class="fa fa-align-justify"></i> <strong>Progress</strong> <small>width</small>
         </div>
         <div>
           <h5>Default width</h5>
@@ -60,7 +60,7 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Progress</strong> <small>height</small>
+          <i class="fa fa-align-justify"></i> <strong>Progress</strong> <small>height</small>
         </div>
         <div>
           <h5>Default height</h5>
@@ -75,7 +75,7 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Progress</strong> <small>variants</small>
+          <i class="fa fa-align-justify"></i> <strong>Progress</strong> <small>variants</small>
         </div>
         <div>
           <div v-for="bar in bars" class="row mb-1">
@@ -93,7 +93,7 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Progress</strong> <small>striped</small>
+          <i class="fa fa-align-justify"></i> <strong>Progress</strong> <small>striped</small>
         </div>
         <div>
           <b-progress :value="25" variant="success" :striped="striped" class="mb-2"></b-progress>
@@ -109,7 +109,7 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Progress</strong> <small>animated</small>
+          <i class="fa fa-align-justify"></i> <strong>Progress</strong> <small>animated</small>
         </div>
         <div>
           <b-progress :value="25" variant="success" striped :animated="animate" class="mb-2"></b-progress>
@@ -125,7 +125,7 @@
         header-tag="header"
         footer-tag="footer">
         <div slot="header">
-          <i className="fa fa-align-justify"></i><strong>Progress</strong> <small>multiple bars</small>
+          <i class="fa fa-align-justify"></i> <strong>Progress</strong> <small>multiple bars</small>
         </div>
         <div>
           <b-progress :max="max3" class="mb-3">